Oss release 2019-07-29
Changes:
- added DcsmRenderer, a smart renderer which uses DCSM to manage
   scene states correctly and asynchronously
- more consistent usage of nullptr
- added documentation for DCSM components
- stability improvements

Original commit sha: 4f8a9c601b47ed5af0eede57d262ccb2e2b1e1e1

+++ external/capu/test/src/util/TraitsTest.cpp
@@ -37,7 +37,7 @@ public:
         val = *&val;
         enum
         {
-            type = ramses_capu::is_CAPU_PRIMITIVE<T>::Value | ramses_capu::is_CAPU_REFERENCE<T>::Value | ramses_capu::is_CAPU_POINTER<T>::Value
+           type = static_cast<uint32_t>(ramses_capu::is_CAPU_PRIMITIVE<T>::Value) | static_cast<uint32_t>(ramses_capu::is_CAPU_REFERENCE<T>::Value) | static_cast<uint32_t>(ramses_capu::is_CAPU_POINTER<T>::Value)
         };
         return type;
     }
